+++ KNUTICE/Presentation/WebView/WebViewController.swift
@@ -66,9 +66,18 @@ final class WebViewController: UIViewController {
         super.viewDidLoad()
         
         navigationItem.largeTitleDisplayMode = .never    //navigation inline title
-        setupNavigationBarItem()
+        view.backgroundColor = .detailViewBackground
+        navigationItem.rightBarButtonItem = UIBarButtonItem(image: UIImage(systemName: "square.and.arrow.up"), style: .plain, target: self, action: #selector(openSharePanel))
         setupLayout()
         loadPage(notice.contentUrl)
+    }
+    
+    private func loadPage(_ url: String) {
+        guard let url = URL(string: url) else {
+            return
+        }
+        
+        webView.load(URLRequest(url: url, cachePolicy: .returnCacheDataElseLoad))
     }
 }
 
@@ -105,6 +114,34 @@ extension WebViewController: WKNavigationDelegate, WKUIDelegate {
         }
         
         return nil
+    }
+}
+
+extension WebViewController {
+    func setupLayout() {
+        //progressView
+        view.addSubview(progressView)
+        progressView.translatesAutoresizingMaskIntoConstraints = false
+        progressView.snp.makeConstraints { make in
+            make.left.top.right.equalTo(self.view.safeAreaLayoutGuide)
+        }
+        
+        //webView
+        view.addSubview(webView)
+        webView.translatesAutoresizingMaskIntoConstraints = false
+        webView.snp.makeConstraints { make in
+            make.top.equalTo(progressView.snp.bottom)
+            make.leading.trailing.bottom.equalToSuperview()
+        }
+        
+        //reminderSheetBtn
+        view.addSubview(reminderSheetBtn)
+        reminderSheetBtn.translatesAutoresizingMaskIntoConstraints = false
+        reminderSheetBtn.snp.makeConstraints { make in
+            make.bottom.equalToSuperview().offset(-50)
+            make.trailing.equalToSuperview().offset(-20)
+            make.width.height.equalTo(50)
+        }
     }
 }
